Updated the onion request interface to be builder-based
